requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
Pillow
python-decouple
whitenoise
gunicorn==21.2.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9